WANN objects encapsulating points/kd trees allowing repeated searches

Methods for WANN objects include

build_tree

delete_tree

getPoints

querySelf

objectPointer

objectPointer

Details

WANN objects will primarily be useful if you make repeated queries. You can also delay building the k-d tree if you are not sure if the object in question will be searched; the tree will automatically be built when it is queried. You can also explicitly control when the tree is built or deleted (for memory management). The tree is wrapped in an R reference class (R5) object which imposes a significant performance penalty for building small trees (< ~ 1000 points). Summarising, WANN objects provide experimental functionality to:

keep ANN points in memory to avoid repeated copying

keep the ANN k-d tree in memory to avoid repeated building

separate building the k-d tree from allocating the points

permit very fast self queries

permit queries of the points from one ANN tree against a second tree

Examples

p1=kcpoints[[1]]
w1=WANN(p1)
# 2 neighbours since the first neighbour will be a self match of each point
w1sq=w1$querySelf(2,0)
w1sq$nn.dists[,2]

#>   [1] 0.9681892 1.0278818 1.0494416 0.9542576 0.8863632 0.9681892 1.0494416
#>   [8] 0.9556508 0.9542576 0.9601247 0.5993720 0.5993720 0.8304804 0.6680217
#>  [15] 0.6680217 0.8304804 0.8643212 0.9556508 1.0187839 0.7144822 0.7144822
#>  [22] 0.6949343 0.6949343 0.9637013 1.1355343 0.5206840 0.5206840 0.5251662
#>  [29] 0.5251662 0.9515838 0.8558675 0.8856917 0.8856917 0.9776708 0.9171703
#>  [36] 0.9171703 0.5198622 0.5198622 0.5232901 1.1061862 1.1914859 0.8558675
#>  [43] 1.0662507 0.8067439 0.8067439 1.1202524 1.2356608 0.8616506 0.8616506
#>  [50] 0.5916783 0.5916783 1.0571831 0.7665660 0.5831901 0.5831901 0.6947572
#>  [57] 0.9097712 0.5489149 0.5489149 0.5311563 0.5311563 0.6682876 0.6682876
#>  [64] 0.6786525 0.7066306 0.6930341 0.6930341 0.8939676 1.3970612 0.6057954
#>  [71] 0.6057954 0.8640905 1.2205859 0.6620545 0.6620545 0.8400811 1.0796623
#>  [78] 0.5085854 0.5085854 0.8400811 0.6714160 0.6714160 0.4874183 0.4840132
#>  [85] 0.4840132 0.7196481 0.8500938 0.8500938 0.4843785 0.4843785 0.9148319
#>  [92] 0.9524522 0.9148319 0.5184018 0.5184018 0.8207075 0.5194473 0.5194473
#>  [99] 1.0200737 0.5866120 0.5859702 0.5859702 0.5885543 0.5885543 0.7728304
#> [106] 0.7552212 0.7728304 0.7656069 1.0591346 1.1517078 0.6231181 0.6231181
#> [113] 0.5334343 0.5334343 0.5236195 0.5236195 0.5152171 0.5152171 0.8032213
#> [120] 1.0200737 0.9342711 1.0116674 1.0269784 0.8200951 0.8200951 1.1185043
#> [127] 1.1375824 0.9657160 0.5845143 0.5845143 0.8722648 0.9786795 0.5847383
#> [134] 0.5847383 0.5547415 0.5547415 1.1174397 0.9164354 0.9164354 1.0028960
#> [141] 0.8765644 0.8765644 0.9700132 0.9700132 0.9935196 0.9935196 1.1374502
#> [148] 1.0073271 0.7443922 0.9201491 0.7443922 0.9201491 0.7444152 1.0721902
#> [155] 0.7430705 0.7430705 0.7465598 0.7496308 0.7490800 0.7490800 0.7889016
#> [162] 0.9645514 0.8059035 0.8059035 0.6823903 0.6823903 0.7950799 0.7950799
#> [169] 0.8044972 0.8044972 0.9346299 0.9203183 0.8373583 0.8341464 0.8341464
#> [176] 0.8579943 0.8579943 0.8749374 1.0288480 0.8764586 0.8764586 0.8690520
#> [183] 0.8690520 1.0438880 1.0073264 0.9830229 0.9830229 0.6113883 0.6113883
#> [190] 0.6115626 0.5945327 0.5945327 0.6139807 0.9590827 0.5820649 0.5820649
#> [197] 0.5916585 0.9992431 0.5731285 0.5731285 0.5846191 0.7718334 1.0888742
#> [204] 0.7487134 0.7487134 0.6268816 0.6268816 0.7578642 1.0044566 0.8519160
#> [211] 0.6745228 0.6745228 0.9924079 0.5949996 0.5949996 0.6076545 0.6052034
#> [218] 0.5942583 0.5780731 0.5780731 0.6149670 0.6149670 0.4468531 0.4468531
#> [225] 0.7109330 1.1058509 1.0044566 0.9924079 0.5453828 0.5442606 0.5442606
#> [232] 0.8225833 0.7947131 0.5124422 0.5124422 0.6520509 0.6520509 0.9119478
#> [239] 0.4843301 0.4776399 0.4776399 0.7045188 0.4692282 0.4692282 0.4818509
#> [246] 0.6841168 0.8663406 0.9414446 0.4986500 0.4986500 0.5208979 0.5208979
#> [253] 0.4734235 0.4734235 0.4799091 0.4799091 0.4973114 0.4973114 0.4981262
#> [260] 0.4981262 0.5005554 0.8710860 0.8405858 0.8405858 0.4998079 0.4998079
#> [267] 0.9430247 0.8484079 0.8907964 0.7191030 0.7191030 0.5845375 0.5845375
#> [274] 0.8524241 0.8249121 0.8249121 1.1744827 0.4831733 0.4831733 1.1106066
#> [281] 0.4726594 0.4621178 0.4621178 0.7704043

w1$query(p1, 2, 0)

#> $nn.idx
#>        [,1] [,2]
#>   [1,]    1    6
#>   [2,]    2    1
#>   [3,]    3    7
#>   [4,]    4    9
#>   [5,]    5   17
#>   [6,]    6    1
#>   [7,]    7    3
#>   [8,]    8   18
#>   [9,]    9    4
#>  [10,]   10   22
#>  [11,]   11   12
#>  [12,]   12   11
#>  [13,]   13   16
#>  [14,]   14   15
#>  [15,]   15   14
#>  [16,]   16   13
#>  [17,]   17   31
#>  [18,]   18    8
#>  [19,]   19   33
#>  [20,]   20   21
#>  [21,]   21   20
#>  [22,]   22   23
#>  [23,]   23   22
#>  [24,]   24   15
#>  [25,]   25   14
#>  [26,]   26   27
#>  [27,]   27   26
#>  [28,]   28   29
#>  [29,]   29   28
#>  [30,]   30   29
#>  [31,]   31   42
#>  [32,]   32   33
#>  [33,]   33   32
#>  [34,]   34   21
#>  [35,]   35   36
#>  [36,]   36   35
#>  [37,]   37   38
#>  [38,]   38   37
#>  [39,]   39   38
#>  [40,]   40   28
#>  [41,]   41   30
#>  [42,]   42   31
#>  [43,]   43   33
#>  [44,]   44   45
#>  [45,]   45   44
#>  [46,]   46   60
#>  [47,]   47   41
#>  [48,]   48   49
#>  [49,]   49   48
#>  [50,]   50   51
#>  [51,]   51   50
#>  [52,]   52   45
#>  [53,]   53   51
#>  [54,]   54   55
#>  [55,]   55   54
#>  [56,]   56   55
#>  [57,]   57   70
#>  [58,]   58   59
#>  [59,]   59   58
#>  [60,]   60   61
#>  [61,]   61   60
#>  [62,]   62   63
#>  [63,]   63   62
#>  [64,]   64   63
#>  [65,]   65   64
#>  [66,]   66   67
#>  [67,]   67   66
#>  [68,]   68   54
#>  [69,]   69   52
#>  [70,]   70   71
#>  [71,]   71   70
#>  [72,]   72   76
#>  [73,]   73   77
#>  [74,]   74   75
#>  [75,]   75   74
#>  [76,]   76   80
#>  [77,]   77   81
#>  [78,]   78   79
#>  [79,]   79   78
#>  [80,]   80   76
#>  [81,]   81   82
#>  [82,]   82   81
#>  [83,]   83   84
#>  [84,]   84   85
#>  [85,]   85   84
#>  [86,]   86   85
#>  [87,]   87   88
#>  [88,]   88   87
#>  [89,]   89   90
#>  [90,]   90   89
#>  [91,]   91   93
#>  [92,]   92  106
#>  [93,]   93   91
#>  [94,]   94   95
#>  [95,]   95   94
#>  [96,]   96   98
#>  [97,]   97   98
#>  [98,]   98   97
#>  [99,]   99  120
#> [100,]  100  101
#> [101,]  101  102
#> [102,]  102  101
#> [103,]  103  104
#> [104,]  104  103
#> [105,]  105  107
#> [106,]  106  104
#> [107,]  107  105
#> [108,]  108  106
#> [109,]  109   93
#> [110,]  110  122
#> [111,]  111  112
#> [112,]  112  111
#> [113,]  113  114
#> [114,]  114  113
#> [115,]  115  116
#> [116,]  116  115
#> [117,]  117  118
#> [118,]  118  117
#> [119,]  119  113
#> [120,]  120   99
#> [121,]  121  107
#> [122,]  122  111
#> [123,]  123  117
#> [124,]  124  125
#> [125,]  125  124
#> [126,]  126  122
#> [127,]  127  123
#> [128,]  128  124
#> [129,]  129  130
#> [130,]  130  129
#> [131,]  131  130
#> [132,]  132  128
#> [133,]  133  134
#> [134,]  134  133
#> [135,]  135  136
#> [136,]  136  135
#> [137,]  137  138
#> [138,]  138  139
#> [139,]  139  138
#> [140,]  140  143
#> [141,]  141  142
#> [142,]  142  141
#> [143,]  143  144
#> [144,]  144  143
#> [145,]  145  146
#> [146,]  146  145
#> [147,]  147  150
#> [148,]  148  146
#> [149,]  149  151
#> [150,]  150  152
#> [151,]  151  149
#> [152,]  152  150
#> [153,]  153  151
#> [154,]  154  155
#> [155,]  155  156
#> [156,]  156  155
#> [157,]  157  156
#> [158,]  158  157
#> [159,]  159  160
#> [160,]  160  159
#> [161,]  161  160
#> [162,]  162  163
#> [163,]  163  164
#> [164,]  164  163
#> [165,]  165  166
#> [166,]  166  165
#> [167,]  167  168
#> [168,]  168  167
#> [169,]  169  170
#> [170,]  170  169
#> [171,]  171  170
#> [172,]  172  173
#> [173,]  173  174
#> [174,]  174  175
#> [175,]  175  174
#> [176,]  176  177
#> [177,]  177  176
#> [178,]  178  177
#> [179,]  179  178
#> [180,]  180  181
#> [181,]  181  180
#> [182,]  182  183
#> [183,]  183  182
#> [184,]  184  185
#> [185,]  185  186
#> [186,]  186  187
#> [187,]  187  186
#> [188,]  188  189
#> [189,]  189  188
#> [190,]  190  189
#> [191,]  191  192
#> [192,]  192  191
#> [193,]  193  192
#> [194,]  194  197
#> [195,]  195  196
#> [196,]  196  195
#> [197,]  197  196
#> [198,]  198  195
#> [199,]  199  200
#> [200,]  200  199
#> [201,]  201  200
#> [202,]  202  199
#> [203,]  203  204
#> [204,]  204  205
#> [205,]  205  204
#> [206,]  206  207
#> [207,]  207  206
#> [208,]  208  207
#> [209,]  209  227
#> [210,]  210  211
#> [211,]  211  212
#> [212,]  212  211
#> [213,]  213  228
#> [214,]  214  215
#> [215,]  215  214
#> [216,]  216  215
#> [217,]  217  218
#> [218,]  218  219
#> [219,]  219  220
#> [220,]  220  219
#> [221,]  221  222
#> [222,]  222  221
#> [223,]  223  224
#> [224,]  224  223
#> [225,]  225  224
#> [226,]  226  237
#> [227,]  227  209
#> [228,]  228  213
#> [229,]  229  230
#> [230,]  230  231
#> [231,]  231  230
#> [232,]  232  231
#> [233,]  233  239
#> [234,]  234  235
#> [235,]  235  234
#> [236,]  236  237
#> [237,]  237  236
#> [238,]  238  232
#> [239,]  239  240
#> [240,]  240  241
#> [241,]  241  240
#> [242,]  242  243
#> [243,]  243  244
#> [244,]  244  243
#> [245,]  245  244
#> [246,]  246  245
#> [247,]  247  251
#> [248,]  248  253
#> [249,]  249  250
#> [250,]  250  249
#> [251,]  251  252
#> [252,]  252  251
#> [253,]  253  254
#> [254,]  254  253
#> [255,]  255  256
#> [256,]  256  255
#> [257,]  257  258
#> [258,]  258  257
#> [259,]  259  260
#> [260,]  260  259
#> [261,]  261  260
#> [262,]  262  263
#> [263,]  263  264
#> [264,]  264  263
#> [265,]  265  266
#> [266,]  266  265
#> [267,]  267  252
#> [268,]  268  257
#> [269,]  269  265
#> [270,]  270  271
#> [271,]  271  270
#> [272,]  272  273
#> [273,]  273  272
#> [274,]  274  275
#> [275,]  275  276
#> [276,]  276  275
#> [277,]  277  280
#> [278,]  278  279
#> [279,]  279  278
#> [280,]  280  284
#> [281,]  281  282
#> [282,]  282  283
#> [283,]  283  282
#> [284,]  284  283
#> 
#> $nn.dists
#>        [,1]      [,2]
#>   [1,]    0 0.9681892
#>   [2,]    0 1.0278818
#>   [3,]    0 1.0494416
#>   [4,]    0 0.9542576
#>   [5,]    0 0.8863632
#>   [6,]    0 0.9681892
#>   [7,]    0 1.0494416
#>   [8,]    0 0.9556508
#>   [9,]    0 0.9542576
#>  [10,]    0 0.9601247
#>  [11,]    0 0.5993720
#>  [12,]    0 0.5993720
#>  [13,]    0 0.8304804
#>  [14,]    0 0.6680217
#>  [15,]    0 0.6680217
#>  [16,]    0 0.8304804
#>  [17,]    0 0.8643212
#>  [18,]    0 0.9556508
#>  [19,]    0 1.0187839
#>  [20,]    0 0.7144822
#>  [21,]    0 0.7144822
#>  [22,]    0 0.6949343
#>  [23,]    0 0.6949343
#>  [24,]    0 0.9637013
#>  [25,]    0 1.1355343
#>  [26,]    0 0.5206840
#>  [27,]    0 0.5206840
#>  [28,]    0 0.5251662
#>  [29,]    0 0.5251662
#>  [30,]    0 0.9515838
#>  [31,]    0 0.8558675
#>  [32,]    0 0.8856917
#>  [33,]    0 0.8856917
#>  [34,]    0 0.9776708
#>  [35,]    0 0.9171703
#>  [36,]    0 0.9171703
#>  [37,]    0 0.5198622
#>  [38,]    0 0.5198622
#>  [39,]    0 0.5232901
#>  [40,]    0 1.1061862
#>  [41,]    0 1.1914859
#>  [42,]    0 0.8558675
#>  [43,]    0 1.0662507
#>  [44,]    0 0.8067439
#>  [45,]    0 0.8067439
#>  [46,]    0 1.1202524
#>  [47,]    0 1.2356608
#>  [48,]    0 0.8616506
#>  [49,]    0 0.8616506
#>  [50,]    0 0.5916783
#>  [51,]    0 0.5916783
#>  [52,]    0 1.0571831
#>  [53,]    0 0.7665660
#>  [54,]    0 0.5831901
#>  [55,]    0 0.5831901
#>  [56,]    0 0.6947572
#>  [57,]    0 0.9097712
#>  [58,]    0 0.5489149
#>  [59,]    0 0.5489149
#>  [60,]    0 0.5311563
#>  [61,]    0 0.5311563
#>  [62,]    0 0.6682876
#>  [63,]    0 0.6682876
#>  [64,]    0 0.6786525
#>  [65,]    0 0.7066306
#>  [66,]    0 0.6930341
#>  [67,]    0 0.6930341
#>  [68,]    0 0.8939676
#>  [69,]    0 1.3970612
#>  [70,]    0 0.6057954
#>  [71,]    0 0.6057954
#>  [72,]    0 0.8640905
#>  [73,]    0 1.2205859
#>  [74,]    0 0.6620545
#>  [75,]    0 0.6620545
#>  [76,]    0 0.8400811
#>  [77,]    0 1.0796623
#>  [78,]    0 0.5085854
#>  [79,]    0 0.5085854
#>  [80,]    0 0.8400811
#>  [81,]    0 0.6714160
#>  [82,]    0 0.6714160
#>  [83,]    0 0.4874183
#>  [84,]    0 0.4840132
#>  [85,]    0 0.4840132
#>  [86,]    0 0.7196481
#>  [87,]    0 0.8500938
#>  [88,]    0 0.8500938
#>  [89,]    0 0.4843785
#>  [90,]    0 0.4843785
#>  [91,]    0 0.9148319
#>  [92,]    0 0.9524522
#>  [93,]    0 0.9148319
#>  [94,]    0 0.5184018
#>  [95,]    0 0.5184018
#>  [96,]    0 0.8207075
#>  [97,]    0 0.5194473
#>  [98,]    0 0.5194473
#>  [99,]    0 1.0200737
#> [100,]    0 0.5866120
#> [101,]    0 0.5859702
#> [102,]    0 0.5859702
#> [103,]    0 0.5885543
#> [104,]    0 0.5885543
#> [105,]    0 0.7728304
#> [106,]    0 0.7552212
#> [107,]    0 0.7728304
#> [108,]    0 0.7656069
#> [109,]    0 1.0591346
#> [110,]    0 1.1517078
#> [111,]    0 0.6231181
#> [112,]    0 0.6231181
#> [113,]    0 0.5334343
#> [114,]    0 0.5334343
#> [115,]    0 0.5236195
#> [116,]    0 0.5236195
#> [117,]    0 0.5152171
#> [118,]    0 0.5152171
#> [119,]    0 0.8032213
#> [120,]    0 1.0200737
#> [121,]    0 0.9342711
#> [122,]    0 1.0116674
#> [123,]    0 1.0269784
#> [124,]    0 0.8200951
#> [125,]    0 0.8200951
#> [126,]    0 1.1185043
#> [127,]    0 1.1375824
#> [128,]    0 0.9657160
#> [129,]    0 0.5845143
#> [130,]    0 0.5845143
#> [131,]    0 0.8722648
#> [132,]    0 0.9786795
#> [133,]    0 0.5847383
#> [134,]    0 0.5847383
#> [135,]    0 0.5547415
#> [136,]    0 0.5547415
#> [137,]    0 1.1174397
#> [138,]    0 0.9164354
#> [139,]    0 0.9164354
#> [140,]    0 1.0028960
#> [141,]    0 0.8765644
#> [142,]    0 0.8765644
#> [143,]    0 0.9700132
#> [144,]    0 0.9700132
#> [145,]    0 0.9935196
#> [146,]    0 0.9935196
#> [147,]    0 1.1374502
#> [148,]    0 1.0073271
#> [149,]    0 0.7443922
#> [150,]    0 0.9201491
#> [151,]    0 0.7443922
#> [152,]    0 0.9201491
#> [153,]    0 0.7444152
#> [154,]    0 1.0721902
#> [155,]    0 0.7430705
#> [156,]    0 0.7430705
#> [157,]    0 0.7465598
#> [158,]    0 0.7496308
#> [159,]    0 0.7490800
#> [160,]    0 0.7490800
#> [161,]    0 0.7889016
#> [162,]    0 0.9645514
#> [163,]    0 0.8059035
#> [164,]    0 0.8059035
#> [165,]    0 0.6823903
#> [166,]    0 0.6823903
#> [167,]    0 0.7950799
#> [168,]    0 0.7950799
#> [169,]    0 0.8044972
#> [170,]    0 0.8044972
#> [171,]    0 0.9346299
#> [172,]    0 0.9203183
#> [173,]    0 0.8373583
#> [174,]    0 0.8341464
#> [175,]    0 0.8341464
#> [176,]    0 0.8579943
#> [177,]    0 0.8579943
#> [178,]    0 0.8749374
#> [179,]    0 1.0288480
#> [180,]    0 0.8764586
#> [181,]    0 0.8764586
#> [182,]    0 0.8690520
#> [183,]    0 0.8690520
#> [184,]    0 1.0438880
#> [185,]    0 1.0073264
#> [186,]    0 0.9830229
#> [187,]    0 0.9830229
#> [188,]    0 0.6113883
#> [189,]    0 0.6113883
#> [190,]    0 0.6115626
#> [191,]    0 0.5945327
#> [192,]    0 0.5945327
#> [193,]    0 0.6139807
#> [194,]    0 0.9590827
#> [195,]    0 0.5820649
#> [196,]    0 0.5820649
#> [197,]    0 0.5916585
#> [198,]    0 0.9992431
#> [199,]    0 0.5731285
#> [200,]    0 0.5731285
#> [201,]    0 0.5846191
#> [202,]    0 0.7718334
#> [203,]    0 1.0888742
#> [204,]    0 0.7487134
#> [205,]    0 0.7487134
#> [206,]    0 0.6268816
#> [207,]    0 0.6268816
#> [208,]    0 0.7578642
#> [209,]    0 1.0044566
#> [210,]    0 0.8519160
#> [211,]    0 0.6745228
#> [212,]    0 0.6745228
#> [213,]    0 0.9924079
#> [214,]    0 0.5949996
#> [215,]    0 0.5949996
#> [216,]    0 0.6076545
#> [217,]    0 0.6052034
#> [218,]    0 0.5942583
#> [219,]    0 0.5780731
#> [220,]    0 0.5780731
#> [221,]    0 0.6149670
#> [222,]    0 0.6149670
#> [223,]    0 0.4468531
#> [224,]    0 0.4468531
#> [225,]    0 0.7109330
#> [226,]    0 1.1058509
#> [227,]    0 1.0044566
#> [228,]    0 0.9924079
#> [229,]    0 0.5453828
#> [230,]    0 0.5442606
#> [231,]    0 0.5442606
#> [232,]    0 0.8225833
#> [233,]    0 0.7947131
#> [234,]    0 0.5124422
#> [235,]    0 0.5124422
#> [236,]    0 0.6520509
#> [237,]    0 0.6520509
#> [238,]    0 0.9119478
#> [239,]    0 0.4843301
#> [240,]    0 0.4776399
#> [241,]    0 0.4776399
#> [242,]    0 0.7045188
#> [243,]    0 0.4692282
#> [244,]    0 0.4692282
#> [245,]    0 0.4818509
#> [246,]    0 0.6841168
#> [247,]    0 0.8663406
#> [248,]    0 0.9414446
#> [249,]    0 0.4986500
#> [250,]    0 0.4986500
#> [251,]    0 0.5208979
#> [252,]    0 0.5208979
#> [253,]    0 0.4734235
#> [254,]    0 0.4734235
#> [255,]    0 0.4799091
#> [256,]    0 0.4799091
#> [257,]    0 0.4973114
#> [258,]    0 0.4973114
#> [259,]    0 0.4981262
#> [260,]    0 0.4981262
#> [261,]    0 0.5005554
#> [262,]    0 0.8710860
#> [263,]    0 0.8405858
#> [264,]    0 0.8405858
#> [265,]    0 0.4998079
#> [266,]    0 0.4998079
#> [267,]    0 0.9430247
#> [268,]    0 0.8484079
#> [269,]    0 0.8907964
#> [270,]    0 0.7191030
#> [271,]    0 0.7191030
#> [272,]    0 0.5845375
#> [273,]    0 0.5845375
#> [274,]    0 0.8524241
#> [275,]    0 0.8249121
#> [276,]    0 0.8249121
#> [277,]    0 1.1744827
#> [278,]    0 0.4831733
#> [279,]    0 0.4831733
#> [280,]    0 1.1106066
#> [281,]    0 0.4726594
#> [282,]    0 0.4621178
#> [283,]    0 0.4621178
#> [284,]    0 0.7704043
#> 

p2=kcpoints[[2]]
w1$query(p2, 1, 0)

#> $nn.idx
#>        [,1]
#>   [1,]   42
#>   [2,]   42
#>   [3,]   42
#>   [4,]   42
#>   [5,]   42
#>   [6,]   42
#>   [7,]    9
#>   [8,]    9
#>   [9,]    9
#>  [10,]    9
#>  [11,]    9
#>  [12,]    9
#>  [13,]    9
#>  [14,]   42
#>  [15,]   42
#>  [16,]   42
#>  [17,]   42
#>  [18,]   42
#>  [19,]   42
#>  [20,]    9
#>  [21,]    9
#>  [22,]    9
#>  [23,]    9
#>  [24,]    9
#>  [25,]   24
#>  [26,]   25
#>  [27,]   42
#>  [28,]   42
#>  [29,]   42
#>  [30,]   42
#>  [31,]   42
#>  [32,]   42
#>  [33,]   12
#>  [34,]   24
#>  [35,]   24
#>  [36,]   15
#>  [37,]   15
#>  [38,]   15
#>  [39,]   25
#>  [40,]   15
#>  [41,]   25
#>  [42,]   25
#>  [43,]   16
#>  [44,]   16
#>  [45,]   42
#>  [46,]   42
#>  [47,]   42
#>  [48,]   13
#>  [49,]   42
#>  [50,]   42
#>  [51,]   42
#>  [52,]   13
#>  [53,]   42
#>  [54,]   42
#>  [55,]   42
#>  [56,]   42
#>  [57,]   15
#>  [58,]   62
#>  [59,]   62
#>  [60,]   62
#>  [61,]   62
#>  [62,]   62
#>  [63,]   62
#>  [64,]   62
#>  [65,]   42
#>  [66,]   13
#>  [67,]   42
#>  [68,]   42
#>  [69,]   13
#>  [70,]   13
#>  [71,]   42
#>  [72,]   13
#>  [73,]   13
#>  [74,]   13
#>  [75,]   42
#>  [76,]   13
#>  [77,]   13
#>  [78,]   13
#>  [79,]   13
#>  [80,]   13
#>  [81,]   13
#>  [82,]   13
#>  [83,]   13
#>  [84,]   13
#>  [85,]   13
#>  [86,]   13
#>  [87,]   63
#>  [88,]   64
#>  [89,]   42
#>  [90,]   42
#>  [91,]   13
#>  [92,]   42
#>  [93,]   13
#>  [94,]   13
#>  [95,]   13
#>  [96,]   13
#>  [97,]   13
#>  [98,]   13
#>  [99,]   13
#> [100,]   13
#> [101,]   64
#> [102,]   64
#> [103,]   67
#> [104,]   42
#> [105,]   13
#> [106,]   13
#> [107,]   13
#> [108,]   13
#> [109,]   13
#> [110,]   13
#> [111,]   67
#> [112,]   67
#> [113,]   13
#> [114,]   13
#> [115,]   13
#> [116,]   13
#> [117,]   13
#> [118,]   67
#> [119,]   67
#> [120,]   67
#> [121,]   13
#> [122,]   13
#> [123,]   13
#> [124,]   13
#> [125,]   13
#> [126,]   13
#> [127,]   13
#> [128,]   67
#> [129,]   67
#> [130,]   13
#> [131,]   13
#> [132,]   13
#> [133,]   13
#> [134,]   67
#> [135,]   67
#> [136,]   67
#> [137,]   13
#> [138,]   13
#> [139,]   13
#> [140,]   13
#> [141,]   13
#> [142,]   13
#> [143,]   13
#> [144,]   13
#> [145,]   67
#> [146,]   42
#> [147,]   13
#> [148,]   13
#> [149,]   13
#> [150,]   13
#> [151,]   13
#> [152,]   13
#> [153,]   67
#> [154,]   67
#> [155,]   13
#> [156,]   67
#> [157,]  115
#> [158,]  115
#> [159,]   67
#> [160,]  115
#> [161,]  115
#> [162,]  115
#> [163,]  115
#> [164,]  115
#> [165,]  115
#> [166,]  115
#> [167,]  115
#> [168,]  115
#> [169,]  115
#> [170,]  127
#> [171,]  127
#> [172,]  127
#> [173,]  127
#> [174,]  127
#> [175,]  127
#> [176,]  147
#> [177,]  149
#> [178,]  149
#> [179,]  149
#> [180,]  149
#> [181,]  149
#> [182,]  149
#> [183,]  151
#> [184,]  153
#> [185,]  153
#> [186,]  153
#> [187,]  158
#> [188,]  158
#> [189,]  158
#> [190,]  161
#> [191,]  161
#> [192,]  161
#> [193,]  164
#> [194,]  164
#> [195,]  164
#> [196,]  168
#> [197,]  168
#> [198,]  172
#> [199,]  175
#> [200,]  175
#> [201,]  175
#> [202,]  178
#> [203,]  178
#> [204,]  178
#> [205,]  178
#> [206,]  178
#> [207,]  178
#> [208,]  183
#> [209,]  183
#> [210,]  183
#> [211,]  183
#> [212,]  183
#> [213,]  190
#> [214,]  190
#> [215,]  190
#> [216,]  190
#> [217,]  190
#> [218,]  190
#> [219,]  190
#> [220,]  190
#> [221,]  190
#> [222,]  190
#> [223,]  190
#> [224,]  190
#> [225,]  190
#> [226,]  190
#> [227,]  190
#> [228,]  190
#> [229,]  190
#> [230,]  190
#> [231,]  234
#> [232,]  234
#> [233,]  234
#> [234,]  234
#> [235,]  234
#> [236,]  235
#> [237,]  235
#> [238,]  235
#> [239,]  235
#> [240,]  235
#> [241,]  235
#> [242,]  235
#> [243,]  235
#> [244,]  235
#> [245,]  235
#> [246,]  234
#> [247,]  235
#> [248,]  234
#> [249,]  246
#> [250,]  246
#> [251,]  246
#> [252,]  246
#> [253,]  243
#> [254,]  246
#> [255,]  264
#> [256,]  264
#> [257,]  264
#> [258,]  264
#> [259,]  243
#> [260,]  243
#> [261,]  268
#> [262,]  264
#> [263,]  264
#> [264,]  264
#> [265,]  264
#> [266,]  268
#> [267,]  268
#> [268,]  268
#> [269,]  268
#> [270,]  268
#> [271,]  268
#> [272,]  268
#> [273,]  268
#> [274,]  268
#> [275,]  268
#> [276,]  268
#> [277,]  268
#> [278,]  268
#> [279,]  268
#> [280,]  268
#> [281,]  268
#> [282,]  268
#> [283,]  264
#> [284,]  264
#> [285,]  264
#> [286,]  264
#> [287,]  264
#> [288,]  283
#> [289,]  283
#> [290,]  283
#> [291,]  283
#> [292,]  264
#> [293,]  264
#> [294,]  283
#> [295,]  283
#> [296,]  283
#> [297,]  283
#> 
#> $nn.dists
#>              [,1]
#>   [1,] 16.3258322
#>   [2,] 10.9198408
#>   [3,] 16.1546249
#>   [4,] 12.0996499
#>   [5,] 10.6541368
#>   [6,] 10.6338308
#>   [7,] 10.1180929
#>   [8,]  9.0083913
#>   [9,]  7.9015184
#>  [10,]  6.2904431
#>  [11,]  3.0950863
#>  [12,]  2.2589226
#>  [13,]  1.8006709
#>  [14,] 15.9632269
#>  [15,] 15.9121524
#>  [16,] 13.9474717
#>  [17,] 13.2860203
#>  [18,] 12.6204052
#>  [19,] 11.1747047
#>  [20,]  6.6226972
#>  [21,]  5.0676497
#>  [22,]  4.4664502
#>  [23,]  3.8768733
#>  [24,]  1.8813893
#>  [25,]  1.1376276
#>  [26,]  2.5738608
#>  [27,] 18.3872601
#>  [28,] 17.7538345
#>  [29,] 17.1223530
#>  [30,] 16.4865638
#>  [31,] 15.1830933
#>  [32,] 14.5359409
#>  [33,]  1.6198515
#>  [34,]  0.9007354
#>  [35,]  0.7453068
#>  [36,]  1.1416242
#>  [37,]  1.5321128
#>  [38,]  2.1943876
#>  [39,]  2.4144708
#>  [40,]  2.6305980
#>  [41,]  3.1178146
#>  [42,]  3.6564636
#>  [43,]  3.8781415
#>  [44,]  4.2056873
#>  [45,] 28.3943679
#>  [46,] 24.5720225
#>  [47,] 23.9258951
#>  [48,] 30.9161034
#>  [49,] 23.1876871
#>  [50,] 22.5249039
#>  [51,] 21.8559666
#>  [52,] 31.8360603
#>  [53,] 21.1152066
#>  [54,] 20.4348741
#>  [55,] 19.7742420
#>  [56,] 19.0660450
#>  [57,]  3.8589228
#>  [58,]  4.3697169
#>  [59,]  3.9376298
#>  [60,]  3.5339019
#>  [61,]  3.1868895
#>  [62,]  3.0988484
#>  [63,]  2.9307684
#>  [64,]  2.8959041
#>  [65,] 27.8904208
#>  [66,] 32.2754059
#>  [67,] 27.2699235
#>  [68,] 26.6616662
#>  [69,] 29.3847647
#>  [70,] 28.8648953
#>  [71,] 25.9976508
#>  [72,] 30.7083487
#>  [73,] 30.1646740
#>  [74,] 29.6391229
#>  [75,] 25.3253647
#>  [76,] 31.3822902
#>  [77,] 32.3732093
#>  [78,] 32.8189286
#>  [79,] 33.9324599
#>  [80,] 33.3029178
#>  [81,] 39.4155038
#>  [82,] 34.4083960
#>  [83,] 36.1428364
#>  [84,] 37.8447769
#>  [85,] 37.2358381
#>  [86,] 36.6104383
#>  [87,]  1.9991341
#>  [88,]  2.2057861
#>  [89,] 27.5209997
#>  [90,] 27.4647923
#>  [91,] 34.5703814
#>  [92,] 27.4316285
#>  [93,] 33.7572097
#>  [94,] 32.9497670
#>  [95,] 31.5979666
#>  [96,] 40.5925093
#>  [97,] 40.4061382
#>  [98,] 38.8195868
#>  [99,] 35.7161907
#> [100,] 35.0997395
#> [101,]  1.7778530
#> [102,]  2.2692876
#> [103,]  2.9852081
#> [104,] 28.2434386
#> [105,] 34.9508321
#> [106,] 42.3864205
#> [107,] 41.5919349
#> [108,] 41.3989523
#> [109,] 41.1906917
#> [110,] 40.9886145
#> [111,]  2.6397967
#> [112,]  3.4259346
#> [113,] 39.4777831
#> [114,] 38.8297413
#> [115,] 36.4003660
#> [116,] 35.7602392
#> [117,] 43.4128173
#> [118,]  3.6270320
#> [119,]  4.1167968
#> [120,]  4.6790190
#> [121,] 40.3707009
#> [122,] 38.1780266
#> [123,] 37.4934637
#> [124,] 46.3566762
#> [125,] 45.7851617
#> [126,] 45.2211215
#> [127,] 44.4519251
#> [128,]  5.7311142
#> [129,]  6.2759203
#> [130,] 42.0678924
#> [131,] 41.2634170
#> [132,] 47.0143243
#> [133,] 47.3943210
#> [134,]  7.2973873
#> [135,]  7.7800348
#> [136,]  8.2795418
#> [137,] 42.9708718
#> [138,] 48.1126838
#> [139,] 48.5263330
#> [140,] 52.8057660
#> [141,] 52.1425006
#> [142,] 51.4966517
#> [143,] 50.0741236
#> [144,] 54.6160782
#> [145,]  9.2196693
#> [146,] 44.6546523
#> [147,] 43.8715817
#> [148,] 49.3392169
#> [149,] 53.7234004
#> [150,] 51.0998389
#> [151,] 49.8098912
#> [152,] 54.2842828
#> [153,]  9.7789049
#> [154,]  9.9371028
#> [155,] 45.5820847
#> [156,] 10.6022248
#> [157,] 10.0292437
#> [158,]  9.5774414
#> [159,] 11.3314676
#> [160,] 10.8624919
#> [161,]  9.3060858
#> [162,]  9.1683688
#> [163,]  9.2882450
#> [164,]  9.0476792
#> [165,]  8.8648378
#> [166,]  8.7168962
#> [167,]  8.6334983
#> [168,]  8.6149151
#> [169,]  8.6512942
#> [170,]  9.0746057
#> [171,] 10.1742702
#> [172,] 11.2927752
#> [173,] 12.0477278
#> [174,] 12.5179912
#> [175,] 13.3722263
#> [176,]  9.8633746
#> [177,] 11.3300483
#> [178,] 10.7563976
#> [179,] 12.9904166
#> [180,] 11.8385513
#> [181,] 11.1454372
#> [182,] 10.8122251
#> [183,]  9.7788330
#> [184,]  9.0193980
#> [185,]  8.6222008
#> [186,]  8.3894430
#> [187,]  7.8376840
#> [188,]  7.1789548
#> [189,]  7.2346690
#> [190,]  6.9853954
#> [191,]  7.1854645
#> [192,]  7.5714252
#> [193,]  7.4414495
#> [194,]  7.8055908
#> [195,]  7.9042366
#> [196,]  7.9326634
#> [197,]  7.8486480
#> [198,]  7.1599538
#> [199,]  6.8029683
#> [200,]  6.9857429
#> [201,]  6.8141053
#> [202,]  6.3022923
#> [203,]  5.8789996
#> [204,]  6.2444776
#> [205,]  6.3516591
#> [206,]  6.4921849
#> [207,]  6.7606984
#> [208,]  7.0931486
#> [209,]  7.7168670
#> [210,]  7.8673427
#> [211,]  8.5778525
#> [212,]  9.4175734
#> [213,] 10.0837556
#> [214,]  9.2332968
#> [215,]  9.7303524
#> [216,]  9.6598581
#> [217,]  9.9388341
#> [218,]  9.9518468
#> [219,] 10.8627590
#> [220,]  8.9768539
#> [221,]  9.0431748
#> [222,] 10.6921694
#> [223,] 11.3949435
#> [224,]  7.9672419
#> [225,]  9.2842098
#> [226,]  8.5019214
#> [227,] 11.6007788
#> [228,]  8.1245824
#> [229,]  8.4597503
#> [230,]  8.8120512
#> [231,] 11.8972011
#> [232,] 11.6747337
#> [233,] 10.5778982
#> [234,]  9.6871081
#> [235,]  9.3349740
#> [236,]  6.8935723
#> [237,]  8.2679879
#> [238,]  8.1993635
#> [239,]  8.1883141
#> [240,]  8.1997669
#> [241,]  6.9907787
#> [242,]  8.3041824
#> [243,]  8.3757500
#> [244,]  8.4615874
#> [245,]  7.6546218
#> [246,]  8.5414807
#> [247,]  8.4967122
#> [248,]  8.7909388
#> [249,]  9.0249475
#> [250,]  7.8319678
#> [251,]  8.3844984
#> [252,]  8.2520978
#> [253,]  7.4392425
#> [254,]  8.2111860
#> [255,]  5.3813050
#> [256,]  5.2836077
#> [257,]  5.9945147
#> [258,]  6.7620094
#> [259,]  7.3292352
#> [260,]  7.8787822
#> [261,]  4.0842482
#> [262,]  4.2187418
#> [263,]  4.3221542
#> [264,]  4.6130254
#> [265,]  5.0459088
#> [266,]  2.8386316
#> [267,]  3.7652995
#> [268,]  3.6200902
#> [269,]  3.7717803
#> [270,]  5.5386158
#> [271,]  5.9977950
#> [272,]  6.4642299
#> [273,]  4.8864718
#> [274,]  7.7479296
#> [275,]  8.2497644
#> [276,]  9.1687997
#> [277,]  9.7074971
#> [278,] 10.2650818
#> [279,] 11.4228222
#> [280,] 11.9988479
#> [281,] 13.0805699
#> [282,] 13.6943742
#> [283,] 14.2345869
#> [284,] 15.1640432
#> [285,] 15.6649828
#> [286,] 16.1616382
#> [287,] 16.6310934
#> [288,] 17.7126954
#> [289,] 17.2383543
#> [290,] 16.7678744
#> [291,] 15.9213584
#> [292,] 17.7282778
#> [293,] 18.0277153
#> [294,] 18.0479274
#> [295,] 15.6411259
#> [296,] 15.1580617
#> [297,] 15.0520411
#> 

# equivalent to
nn2(p1, p2, k=1)

#> $nn.idx
#>        [,1]
#>   [1,]   42
#>   [2,]   42
#>   [3,]   42
#>   [4,]   42
#>   [5,]   42
#>   [6,]   42
#>   [7,]    9
#>   [8,]    9
#>   [9,]    9
#>  [10,]    9
#>  [11,]    9
#>  [12,]    9
#>  [13,]    9
#>  [14,]   42
#>  [15,]   42
#>  [16,]   42
#>  [17,]   42
#>  [18,]   42
#>  [19,]   42
#>  [20,]    9
#>  [21,]    9
#>  [22,]    9
#>  [23,]    9
#>  [24,]    9
#>  [25,]   24
#>  [26,]   25
#>  [27,]   42
#>  [28,]   42
#>  [29,]   42
#>  [30,]   42
#>  [31,]   42
#>  [32,]   42
#>  [33,]   12
#>  [34,]   24
#>  [35,]   24
#>  [36,]   15
#>  [37,]   15
#>  [38,]   15
#>  [39,]   25
#>  [40,]   15
#>  [41,]   25
#>  [42,]   25
#>  [43,]   16
#>  [44,]   16
#>  [45,]   42
#>  [46,]   42
#>  [47,]   42
#>  [48,]   13
#>  [49,]   42
#>  [50,]   42
#>  [51,]   42
#>  [52,]   13
#>  [53,]   42
#>  [54,]   42
#>  [55,]   42
#>  [56,]   42
#>  [57,]   15
#>  [58,]   62
#>  [59,]   62
#>  [60,]   62
#>  [61,]   62
#>  [62,]   62
#>  [63,]   62
#>  [64,]   62
#>  [65,]   42
#>  [66,]   13
#>  [67,]   42
#>  [68,]   42
#>  [69,]   13
#>  [70,]   13
#>  [71,]   42
#>  [72,]   13
#>  [73,]   13
#>  [74,]   13
#>  [75,]   42
#>  [76,]   13
#>  [77,]   13
#>  [78,]   13
#>  [79,]   13
#>  [80,]   13
#>  [81,]   13
#>  [82,]   13
#>  [83,]   13
#>  [84,]   13
#>  [85,]   13
#>  [86,]   13
#>  [87,]   63
#>  [88,]   64
#>  [89,]   42
#>  [90,]   42
#>  [91,]   13
#>  [92,]   42
#>  [93,]   13
#>  [94,]   13
#>  [95,]   13
#>  [96,]   13
#>  [97,]   13
#>  [98,]   13
#>  [99,]   13
#> [100,]   13
#> [101,]   64
#> [102,]   64
#> [103,]   67
#> [104,]   42
#> [105,]   13
#> [106,]   13
#> [107,]   13
#> [108,]   13
#> [109,]   13
#> [110,]   13
#> [111,]   67
#> [112,]   67
#> [113,]   13
#> [114,]   13
#> [115,]   13
#> [116,]   13
#> [117,]   13
#> [118,]   67
#> [119,]   67
#> [120,]   67
#> [121,]   13
#> [122,]   13
#> [123,]   13
#> [124,]   13
#> [125,]   13
#> [126,]   13
#> [127,]   13
#> [128,]   67
#> [129,]   67
#> [130,]   13
#> [131,]   13
#> [132,]   13
#> [133,]   13
#> [134,]   67
#> [135,]   67
#> [136,]   67
#> [137,]   13
#> [138,]   13
#> [139,]   13
#> [140,]   13
#> [141,]   13
#> [142,]   13
#> [143,]   13
#> [144,]   13
#> [145,]   67
#> [146,]   42
#> [147,]   13
#> [148,]   13
#> [149,]   13
#> [150,]   13
#> [151,]   13
#> [152,]   13
#> [153,]   67
#> [154,]   67
#> [155,]   13
#> [156,]   67
#> [157,]  115
#> [158,]  115
#> [159,]   67
#> [160,]  115
#> [161,]  115
#> [162,]  115
#> [163,]  115
#> [164,]  115
#> [165,]  115
#> [166,]  115
#> [167,]  115
#> [168,]  115
#> [169,]  115
#> [170,]  127
#> [171,]  127
#> [172,]  127
#> [173,]  127
#> [174,]  127
#> [175,]  127
#> [176,]  147
#> [177,]  149
#> [178,]  149
#> [179,]  149
#> [180,]  149
#> [181,]  149
#> [182,]  149
#> [183,]  151
#> [184,]  153
#> [185,]  153
#> [186,]  153
#> [187,]  158
#> [188,]  158
#> [189,]  158
#> [190,]  161
#> [191,]  161
#> [192,]  161
#> [193,]  164
#> [194,]  164
#> [195,]  164
#> [196,]  168
#> [197,]  168
#> [198,]  172
#> [199,]  175
#> [200,]  175
#> [201,]  175
#> [202,]  178
#> [203,]  178
#> [204,]  178
#> [205,]  178
#> [206,]  178
#> [207,]  178
#> [208,]  183
#> [209,]  183
#> [210,]  183
#> [211,]  183
#> [212,]  183
#> [213,]  190
#> [214,]  190
#> [215,]  190
#> [216,]  190
#> [217,]  190
#> [218,]  190
#> [219,]  190
#> [220,]  190
#> [221,]  190
#> [222,]  190
#> [223,]  190
#> [224,]  190
#> [225,]  190
#> [226,]  190
#> [227,]  190
#> [228,]  190
#> [229,]  190
#> [230,]  190
#> [231,]  234
#> [232,]  234
#> [233,]  234
#> [234,]  234
#> [235,]  234
#> [236,]  235
#> [237,]  235
#> [238,]  235
#> [239,]  235
#> [240,]  235
#> [241,]  235
#> [242,]  235
#> [243,]  235
#> [244,]  235
#> [245,]  235
#> [246,]  234
#> [247,]  235
#> [248,]  234
#> [249,]  246
#> [250,]  246
#> [251,]  246
#> [252,]  246
#> [253,]  243
#> [254,]  246
#> [255,]  264
#> [256,]  264
#> [257,]  264
#> [258,]  264
#> [259,]  243
#> [260,]  243
#> [261,]  268
#> [262,]  264
#> [263,]  264
#> [264,]  264
#> [265,]  264
#> [266,]  268
#> [267,]  268
#> [268,]  268
#> [269,]  268
#> [270,]  268
#> [271,]  268
#> [272,]  268
#> [273,]  268
#> [274,]  268
#> [275,]  268
#> [276,]  268
#> [277,]  268
#> [278,]  268
#> [279,]  268
#> [280,]  268
#> [281,]  268
#> [282,]  268
#> [283,]  264
#> [284,]  264
#> [285,]  264
#> [286,]  264
#> [287,]  264
#> [288,]  283
#> [289,]  283
#> [290,]  283
#> [291,]  283
#> [292,]  264
#> [293,]  264
#> [294,]  283
#> [295,]  283
#> [296,]  283
#> [297,]  283
#> 
#> $nn.dists
#>              [,1]
#>   [1,] 16.3258322
#>   [2,] 10.9198408
#>   [3,] 16.1546249
#>   [4,] 12.0996499
#>   [5,] 10.6541368
#>   [6,] 10.6338308
#>   [7,] 10.1180929
#>   [8,]  9.0083913
#>   [9,]  7.9015184
#>  [10,]  6.2904431
#>  [11,]  3.0950863
#>  [12,]  2.2589226
#>  [13,]  1.8006709
#>  [14,] 15.9632269
#>  [15,] 15.9121524
#>  [16,] 13.9474717
#>  [17,] 13.2860203
#>  [18,] 12.6204052
#>  [19,] 11.1747047
#>  [20,]  6.6226972
#>  [21,]  5.0676497
#>  [22,]  4.4664502
#>  [23,]  3.8768733
#>  [24,]  1.8813893
#>  [25,]  1.1376276
#>  [26,]  2.5738608
#>  [27,] 18.3872601
#>  [28,] 17.7538345
#>  [29,] 17.1223530
#>  [30,] 16.4865638
#>  [31,] 15.1830933
#>  [32,] 14.5359409
#>  [33,]  1.6198515
#>  [34,]  0.9007354
#>  [35,]  0.7453068
#>  [36,]  1.1416242
#>  [37,]  1.5321128
#>  [38,]  2.1943876
#>  [39,]  2.4144708
#>  [40,]  2.6305980
#>  [41,]  3.1178146
#>  [42,]  3.6564636
#>  [43,]  3.8781415
#>  [44,]  4.2056873
#>  [45,] 28.3943679
#>  [46,] 24.5720225
#>  [47,] 23.9258951
#>  [48,] 30.9161034
#>  [49,] 23.1876871
#>  [50,] 22.5249039
#>  [51,] 21.8559666
#>  [52,] 31.8360603
#>  [53,] 21.1152066
#>  [54,] 20.4348741
#>  [55,] 19.7742420
#>  [56,] 19.0660450
#>  [57,]  3.8589228
#>  [58,]  4.3697169
#>  [59,]  3.9376298
#>  [60,]  3.5339019
#>  [61,]  3.1868895
#>  [62,]  3.0988484
#>  [63,]  2.9307684
#>  [64,]  2.8959041
#>  [65,] 27.8904208
#>  [66,] 32.2754059
#>  [67,] 27.2699235
#>  [68,] 26.6616662
#>  [69,] 29.3847647
#>  [70,] 28.8648953
#>  [71,] 25.9976508
#>  [72,] 30.7083487
#>  [73,] 30.1646740
#>  [74,] 29.6391229
#>  [75,] 25.3253647
#>  [76,] 31.3822902
#>  [77,] 32.3732093
#>  [78,] 32.8189286
#>  [79,] 33.9324599
#>  [80,] 33.3029178
#>  [81,] 39.4155038
#>  [82,] 34.4083960
#>  [83,] 36.1428364
#>  [84,] 37.8447769
#>  [85,] 37.2358381
#>  [86,] 36.6104383
#>  [87,]  1.9991341
#>  [88,]  2.2057861
#>  [89,] 27.5209997
#>  [90,] 27.4647923
#>  [91,] 34.5703814
#>  [92,] 27.4316285
#>  [93,] 33.7572097
#>  [94,] 32.9497670
#>  [95,] 31.5979666
#>  [96,] 40.5925093
#>  [97,] 40.4061382
#>  [98,] 38.8195868
#>  [99,] 35.7161907
#> [100,] 35.0997395
#> [101,]  1.7778530
#> [102,]  2.2692876
#> [103,]  2.9852081
#> [104,] 28.2434386
#> [105,] 34.9508321
#> [106,] 42.3864205
#> [107,] 41.5919349
#> [108,] 41.3989523
#> [109,] 41.1906917
#> [110,] 40.9886145
#> [111,]  2.6397967
#> [112,]  3.4259346
#> [113,] 39.4777831
#> [114,] 38.8297413
#> [115,] 36.4003660
#> [116,] 35.7602392
#> [117,] 43.4128173
#> [118,]  3.6270320
#> [119,]  4.1167968
#> [120,]  4.6790190
#> [121,] 40.3707009
#> [122,] 38.1780266
#> [123,] 37.4934637
#> [124,] 46.3566762
#> [125,] 45.7851617
#> [126,] 45.2211215
#> [127,] 44.4519251
#> [128,]  5.7311142
#> [129,]  6.2759203
#> [130,] 42.0678924
#> [131,] 41.2634170
#> [132,] 47.0143243
#> [133,] 47.3943210
#> [134,]  7.2973873
#> [135,]  7.7800348
#> [136,]  8.2795418
#> [137,] 42.9708718
#> [138,] 48.1126838
#> [139,] 48.5263330
#> [140,] 52.8057660
#> [141,] 52.1425006
#> [142,] 51.4966517
#> [143,] 50.0741236
#> [144,] 54.6160782
#> [145,]  9.2196693
#> [146,] 44.6546523
#> [147,] 43.8715817
#> [148,] 49.3392169
#> [149,] 53.7234004
#> [150,] 51.0998389
#> [151,] 49.8098912
#> [152,] 54.2842828
#> [153,]  9.7789049
#> [154,]  9.9371028
#> [155,] 45.5820847
#> [156,] 10.6022248
#> [157,] 10.0292437
#> [158,]  9.5774414
#> [159,] 11.3314676
#> [160,] 10.8624919
#> [161,]  9.3060858
#> [162,]  9.1683688
#> [163,]  9.2882450
#> [164,]  9.0476792
#> [165,]  8.8648378
#> [166,]  8.7168962
#> [167,]  8.6334983
#> [168,]  8.6149151
#> [169,]  8.6512942
#> [170,]  9.0746057
#> [171,] 10.1742702
#> [172,] 11.2927752
#> [173,] 12.0477278
#> [174,] 12.5179912
#> [175,] 13.3722263
#> [176,]  9.8633746
#> [177,] 11.3300483
#> [178,] 10.7563976
#> [179,] 12.9904166
#> [180,] 11.8385513
#> [181,] 11.1454372
#> [182,] 10.8122251
#> [183,]  9.7788330
#> [184,]  9.0193980
#> [185,]  8.6222008
#> [186,]  8.3894430
#> [187,]  7.8376840
#> [188,]  7.1789548
#> [189,]  7.2346690
#> [190,]  6.9853954
#> [191,]  7.1854645
#> [192,]  7.5714252
#> [193,]  7.4414495
#> [194,]  7.8055908
#> [195,]  7.9042366
#> [196,]  7.9326634
#> [197,]  7.8486480
#> [198,]  7.1599538
#> [199,]  6.8029683
#> [200,]  6.9857429
#> [201,]  6.8141053
#> [202,]  6.3022923
#> [203,]  5.8789996
#> [204,]  6.2444776
#> [205,]  6.3516591
#> [206,]  6.4921849
#> [207,]  6.7606984
#> [208,]  7.0931486
#> [209,]  7.7168670
#> [210,]  7.8673427
#> [211,]  8.5778525
#> [212,]  9.4175734
#> [213,] 10.0837556
#> [214,]  9.2332968
#> [215,]  9.7303524
#> [216,]  9.6598581
#> [217,]  9.9388341
#> [218,]  9.9518468
#> [219,] 10.8627590
#> [220,]  8.9768539
#> [221,]  9.0431748
#> [222,] 10.6921694
#> [223,] 11.3949435
#> [224,]  7.9672419
#> [225,]  9.2842098
#> [226,]  8.5019214
#> [227,] 11.6007788
#> [228,]  8.1245824
#> [229,]  8.4597503
#> [230,]  8.8120512
#> [231,] 11.8972011
#> [232,] 11.6747337
#> [233,] 10.5778982
#> [234,]  9.6871081
#> [235,]  9.3349740
#> [236,]  6.8935723
#> [237,]  8.2679879
#> [238,]  8.1993635
#> [239,]  8.1883141
#> [240,]  8.1997669
#> [241,]  6.9907787
#> [242,]  8.3041824
#> [243,]  8.3757500
#> [244,]  8.4615874
#> [245,]  7.6546218
#> [246,]  8.5414807
#> [247,]  8.4967122
#> [248,]  8.7909388
#> [249,]  9.0249475
#> [250,]  7.8319678
#> [251,]  8.3844984
#> [252,]  8.2520978
#> [253,]  7.4392425
#> [254,]  8.2111860
#> [255,]  5.3813050
#> [256,]  5.2836077
#> [257,]  5.9945147
#> [258,]  6.7620094
#> [259,]  7.3292352
#> [260,]  7.8787822
#> [261,]  4.0842482
#> [262,]  4.2187418
#> [263,]  4.3221542
#> [264,]  4.6130254
#> [265,]  5.0459088
#> [266,]  2.8386316
#> [267,]  3.7652995
#> [268,]  3.6200902
#> [269,]  3.7717803
#> [270,]  5.5386158
#> [271,]  5.9977950
#> [272,]  6.4642299
#> [273,]  4.8864718
#> [274,]  7.7479296
#> [275,]  8.2497644
#> [276,]  9.1687997
#> [277,]  9.7074971
#> [278,] 10.2650818
#> [279,] 11.4228222
#> [280,] 11.9988479
#> [281,] 13.0805699
#> [282,] 13.6943742
#> [283,] 14.2345869
#> [284,] 15.1640432
#> [285,] 15.6649828
#> [286,] 16.1616382
#> [287,] 16.6310934
#> [288,] 17.7126954
#> [289,] 17.2383543
#> [290,] 16.7678744
#> [291,] 15.9213584
#> [292,] 17.7282778
#> [293,] 18.0277153
#> [294,] 18.0479274
#> [295,] 15.6411259
#> [296,] 15.1580617
#> [297,] 15.0520411
#> 

# Explicitly build or delete k-d tree
w1$build_tree()
w1$delete_tree()